Fix not checking user-entered values correctly
The getItemAtPosition() function returns an array element from the
original puzzle. However, that array is never modified, since when
the text is changed through the dialog, that only affects the
TextView and not the puzzle array.
Therefore, to actually check the value shown on screen, we needed
to get the content of the TextView. This commit adds the function
`getDisplayedText(int position)` to do that.